requirements.txt
tweepy==4.14.0
python-dotenv==1.0.0
google-generativeai==0.3.1
groq==0.13.1
pillow==10.0.0
schedule==1.2.0
requests==2.31.0
firebase-admin>=6.5.0 